refactor(avio): relocate the editing model into the engine crate
Move the editing model up out of the primitive crates into the top-level avio
engine so the ff-* family becomes model-free by dependency direction.

- Move Timeline / TimelineBuilder / Clip / VideoEffectRenderer and the
  Timeline::render derivation from ff-pipeline into avio (git rename; the bodies
  are unchanged). The model now returns a new avio::TimelineError, and
  ff_pipeline::PipelineError drops its two model variants (TimelineRenderFailed,
  ClipNotFound); the execution pipelines keep the rest.
- Move the Timeline -> Scene projection into avio::Timeline::to_scene and add
  avio::TimelinePlayer, which derives a Scene and hands it to ff-preview. The
  real-time runner stays in ff-preview: its former TimelinePlayer becomes
  ScenePlayer (a Scene consumer), and the crate drops its ff-pipeline dependency
  from the timeline feature (still optional for proxy).
- ff-pipeline exposes hwaccel_to_hardware_encoder for the relocated render, and
  drops the now-dead ff_filter enum re-export.
- The avio facade defines these types instead of re-exporting them from
  ff-pipeline; relocate the model's integration tests (and their fixtures) into
  avio, gated on the pipeline feature.

Behaviour-preserving: verified by the relocated integration tests, a to_scene
projection unit test, and the avio-examples end-to-end harness.
